moved ingame message folder to better loccation. Added scroll to bottom to history game moves
