+++ src/components/Dashboard/GameChat/GameChat.css
@@ -62,7 +62,6 @@
     opacity: 1;
 }
 
-
 .message-entry img{
     height: 20px;
     width: 20px;
