+++ src/components/Dashboard/GameChat/MessageIcon/MessageIconTypes.ts
@@ -1,4 +1,3 @@
-
 export interface MessageIconProps {
     messageBadge: number;
     messageStyle: string;
